implement Exercise6
