+++ src/hooks/useHandleKeyDown.js
@@ -1,0 +1,10 @@
+import { useEffect } from 'react';
+
+export function useHandleKeyDown({ handleKeyDown }) {
+  useEffect(() => {
+    document.addEventListener('keydown', handleKeyDown);
+    return () => {
+      document.removeEventListener('keydown', handleKeyDown);
+    };
+  }, [handleKeyDown]);
+}
